# src/docs/swagger/wizard.service.json
{"swagger":"2.0","info":{"version":"v1","title":"Wizard Service"},"basePath":"/","paths":{"/v1/chargetype":{"post":{"tags":["ChargeType"],"operationId":"V1ChargetypePost","consumes":["application/json","text/json","application/json-patch+json"],"produces":[],"parameters":[{"name":"request","in":"body","required":false,"schema":{"$ref":"#/definitions/PostChargeTypeRequest"}}],"responses":{"200":{"description":"Success"}}}},"/v1/chargetype/{email}":{"get":{"tags":["ChargeType"],"operationId":"V1ChargetypeByEmailGet","consumes":[],"produces":[],"parameters":[{"name":"email","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/v1/financial":{"post":{"tags":["Financial"],"operationId":"V1FinancialPost","consumes":["application/json","text/json","application/json-patch+json"],"produces":[],"parameters":[{"name":"request","in":"body","required":false,"schema":{"$ref":"#/definitions/PostFinancialRequest"}}],"responses":{"200":{"description":"Success"}}}},"/v1/financial/{email}":{"get":{"tags":["Financial"],"operationId":"V1FinancialByEmailGet","consumes":[],"produces":[],"parameters":[{"name":"email","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/healthcheck":{"get":{"tags":["Healthcheck"],"operationId":"HealthcheckGet","consumes":[],"produces":[],"responses":{"200":{"description":"Success"}}}},"/ping":{"get":{"tags":["Healthcheck"],"operationId":"PingGet","consumes":[],"produces":[],"responses":{"200":{"description":"Success"}}}},"/v1/horse":{"post":{"tags":["Horse"],"operationId":"V1HorsePost","consumes":["application/json","text/json","application/json-patch+json"],"produces":[],"parameters":[{"name":"request","in":"body","required":false,"schema":{"$ref":"#/definitions/PostChargeTypeRequest"}}],"responses":{"200":{"description":"Success"}}}},"/v1/horse/{email}":{"get":{"tags":["Horse"],"operationId":"V1HorseByEmailGet","consumes":[],"produces":[],"parameters":[{"name":"email","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/v1/stable":{"post":{"tags":["Stable"],"operationId":"V1StablePost","consumes":["application/json","text/json","application/json-patch+json"],"produces":[],"parameters":[{"name":"request","in":"body","required":false,"schema":{"$ref":"#/definitions/PostStableDataRequest"}}],"responses":{"200":{"description":"Success"}}}},"/v1/stable/{email}":{"get":{"tags":["Stable"],"operationId":"V1StableByEmailGet","consumes":[],"produces":[],"parameters":[{"name":"email","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}}}},"definitions":{"PostChargeTypeRequest":{"type":"object","properties":{"stableChargeTypes":{"type":"array","items":{"$ref":"#/definitions/StableChargeType"}},"standardChargeTypes":{"type":"array","items":{"$ref":"#/definitions/StandardChargeType"}}}},"StableChargeType":{"type":"object","properties":{"id":{"format":"int64","type":"integer"},"description":{"type":"string"},"rate":{"type":"string"}}},"StandardChargeType":{"type":"object","properties":{"id":{"format":"int64","type":"integer"},"description":{"type":"string"},"rate":{"type":"string"},"inStable":{"type":"boolean"},"unit":{"type":"string"}}},"PostFinancialRequest":{"type":"object","properties":{"taxNumber":{"type":"string"},"taxRate":{"type":"string"},"taxDate":{"type":"string"},"previousTaxRate":{"type":"string"},"nextInvoiceNumber":{"type":"string"}}},"PostStableDataRequest":{"type":"object","properties":{"racingCode":{"type":"string"},"trainer":{"type":"string"},"legalEntity":{"type":"string"},"mobile":{"type":"string"},"phone":{"type":"string"},"fax":{"type":"string"},"address":{"type":"string"}}}},"securityDefinitions":{}}